# pluto/api/swagger.json
{
   "swagger": "2.0",
   "info": {
      "description": "",
      "version": "1.0.0",
      "title": "pluto"
   },
...
   "paths": {
      "/users/login": {
         "post": {
            "tags": [
               "users"
            ],
            "summary": "User Login",
            "description": "User Login",
            "consumes": [
               "application/json"
            ],
            "produces": [
               "application/json"
            ],
            "parameters": [
               {
                  "in": "body",
                  "name": "args",
                  "description": "username, password",
                  "required": true,
                  "schema": {
                     "$ref": "#/definitions/UserLogin"
                  }
               }
            ],
            "responses": {
               "405": {
                  "description": "Invalid input"
               }
            }
         }
      },
...
   },
   "securityDefinitions": {
      "api_key": {
         "type": "apiKey",
         "in": "header",
         "name": "x-access-token"
      }
   },
   "definitions": {
      "UserLogin": {
         "properties": {
            "email": {
               "type": "string"
            },
            "password": {
               "type": "string"
            }
         }
      },
...
   }
}